refactor(server): move inline tests to sibling tests.rs files (batch 1)
Promote the inline #[cfg(test)] mod tests blocks in audiobooks.rs,
author_photos.rs, and ebooks.rs to sibling tests.rs files following the
db/src/books/tests.rs pattern called out in 03-unit-testing.md and
05-rust-style.md. Pure mechanical move — same tests, same fixtures, no
behavior change.

Production line counts after:
- audiobooks.rs: 1141 -> 351 (+ audiobooks/tests.rs at 793)
- author_photos.rs: 1000 -> 318 (+ author_photos/tests.rs at 682)
- ebooks.rs: 602 -> 94 (+ ebooks/tests.rs at 511)

cargo test -p omnibus before: 185 passed
cargo test -p omnibus after:  185 passed

Refs #376
